==> 5.MNLI_fb.ipynb <==
from transformers import AutoTokenizer, pipeline
from transformers.pipelines.pt_utils import KeyDataset
import datasets, pandas as pd

target = "MoritzLaurer/DeBERTa-v3-base-mnli-fever-anli"
tokenizer = AutoTokenizer.from_pretrained(target)

hypothesis_template = "The stance of this comment {}."
template = {
    "Favor, Against":"favors Trump and is against Clinton",
    "Favor, None":"favors Trump and does not express any stance toward Clinton",
    "Against, Favor":"is against Trump and favors Clinton",
    "Against, Against":"is against Trump and against Clinton",
    "Against, None":"is against Trump and does not express any stance toward Clinton",
    "None, Favor":"does not express any stance toward Trump and favors Clinton",
    "None, Against":"does not express any stance toward Trump and is against Clinton",
    "None, None":"does not express stance toward Trump nor Clinton"
    }

id2label = pd.DataFrame.from_dict(template, orient="index").iloc[:,0]
label2id = {v: k for k, v in id2label.items()}

test = pd.read_csv("data/fb_test.csv")
test_ds = datasets.Dataset.from_pandas(test)

pipe = pipeline('zero-shot-classification',
                model=target,
                tokenizer=tokenizer)

preds = []
for out in pipe(KeyDataset(test_ds, "prompt"), id2label.tolist(),
        hypothesis_template=hypothesis_template,
        multi_label=False,
        batch_size=16):
        preds.append(out["labels"][0])
test_preds = pd.concat([test, pd.DataFrame({"preds":preds}).replace(label2id, regex=True)], axis=1)
test_preds.to_csv(f"predicted_labels/fb_deberta_mnli.csv", index=False)
